Added the abilit to register a device token with Urban Airship
This can be useful for new devices and when a device token becomes inactive within UA system.

Created an new DeviceToken class, PutRequest and DeviceTokenRequest and extended the UrbanAirSharpGatway to expose this in a simple method call.

Also added a test in the TestApp so you can register a device token